Fixes #28 . App crash (pick image/images from GDrive without Crop option)
