+++ rx_paparazzo/src/main/java/com/fuck_boilerplate/rx_paparazzo/internal/di/ApplicationComponentImpl.java
@@ -4,6 +4,8 @@ package com.fuck_boilerplate.rx_paparazzo.internal.di;
 import com.fuck_boilerplate.rx_paparazzo.entities.Config;
 import com.fuck_boilerplate.rx_paparazzo.entities.TargetUi;
 import com.fuck_boilerplate.rx_paparazzo.interactors.CropImage;
+import com.fuck_boilerplate.rx_paparazzo.interactors.DownloadImage;
+import com.fuck_boilerplate.rx_paparazzo.interactors.ImageUtils;
 import com.fuck_boilerplate.rx_paparazzo.interactors.GetDimens;
 import com.fuck_boilerplate.rx_paparazzo.interactors.GetPath;
 import com.fuck_boilerplate.rx_paparazzo.interactors.GrantPermissions;
@@ -16,6 +18,8 @@ import com.fuck_boilerplate.rx_paparazzo.workers.Camera;
 import com.fuck_boilerplate.rx_paparazzo.workers.Gallery;
 
 class ApplicationComponentImpl extends ApplicationComponent {
+    private final ImageUtils imageUtils;
+    private final DownloadImage downloadImage;
     private final StartIntent startIntent;
     private final GetPath getPath;
     private final GetDimens getDimens;
@@ -30,11 +34,13 @@ class ApplicationComponentImpl extends ApplicationComponent {
 
     public ApplicationComponentImpl(TargetUi ui, Config config) {
         startIntent = new StartIntent(ui);
-        getPath = new GetPath(ui);
+        imageUtils = new ImageUtils(ui, config);
+        downloadImage = new DownloadImage(ui, imageUtils);
+        getPath = new GetPath(ui, imageUtils, downloadImage);
         takePhoto = new TakePhoto(startIntent, ui);
         getDimens = new GetDimens(ui, config, getPath);
         cropImage = new CropImage(ui, config, startIntent, getPath, getDimens);
-        saveImage = new SaveImage(ui, config, getPath, getDimens);
+        saveImage = new SaveImage(ui, config, getPath, getDimens, imageUtils);
         grantPermissions = new GrantPermissions(ui);
         pickImages = new PickImages(startIntent);
         pickImage = new PickImage(startIntent, ui);
